Enum cairo_sys::enums::Antialias

pub enum Antialias {
    Default,
    None,
    Gray,
    Subpixel,
    Fast,
    Good,
    Best,
}

Specifies the type of antialiasing to do when rendering text or shapes.

As it is not necessarily clear from the above what advantages a particular antialias method provides, since 1.12, there is also a set of hints: Fast: Allow the backend to degrade raster quality for speed. Good: A balance between speed and quality. Best: A high-fidelity, but potentially slow, raster mode.

These make no guarantee on how the backend will perform its rasterisation (if it even rasterises!), nor that they have any differing effect other than to enable some form of antialiasing. In the case of glyph rendering, Fast and Good will be mapped to Gray, with Best being equivalent to Subpixel.

The interpretation of Default is left entirely up to the backend, typically this will be similar to Good.

Variants

Default

Use the default antialiasing for the subsystem and target device, since 1.0

None

Use a bilevel alpha mask, since 1.0

Gray

Perform single-color antialiasing (using shades of gray for black text on a white background, for example), since 1.0

Subpixel

Perform antialiasing by taking advantage of the order of subpixel elements on devices such as LCD panels, since 1.0

Fast

Hint that the backend should perform some antialiasing but prefer speed over quality, since 1.12

Good

The backend should balance quality against performance, since 1.12

Best

Hint that the backend should render at the highest quality, sacrificing speed if necessary, since 1.12